+++ pytorch_lightning/plugins/environments/cluster_environment.py
@@ -13,13 +13,29 @@
 # limitations under the License.
 from abc import ABC, abstractmethod
 
+from pytorch_lightning.utilities import rank_zero_deprecation
+
 
 class ClusterEnvironment(ABC):
     """Specification of a cluster environment."""
 
+    @property
     @abstractmethod
-    def creates_children(self) -> bool:
+    def creates_processes_externally(self) -> bool:
         """Whether the environment creates the subprocesses or not."""
+
+    def creates_children(self) -> bool:
+        """Whether the environment creates the subprocesses or not.
+
+        .. deprecated:: v1.5
+            This method was deprecated in v1.5 and will be removed in v1.6. Use the property
+            :attr:`creates_processes_externally` instead.
+        """
+        rank_zero_deprecation(
+            f"`{self.__class__.__name__}.creates_children()` was deprecated in v1.5 and will be removed in v1.6."
+            " Use the property :attr:`creates_processes_externally` instead."
+        )
+        return self.creates_processes_externally
 
     @abstractmethod
     def master_address(self) -> str: